cloud_function/requirements.txt
google-cloud-bigquery
google-cloud-secret-manager
simple-salesforce
lxml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask
gunicorn
